requirements.txt
# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Trading API
alpaca-trade-api==3.1.1
websocket-client==1.6.3

# Alternative Data Sources (for free market data)
yfinance==0.2.22
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing and Analytics
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<1.27.0
matplotlib>=3.7.0,<3.9.0
scipy>=1.10.0,<1.12.0
plotly==5.17.0

# Date/Time Handling
pytz==2023.3.post1
python-dateutil==2.8.2

# Configuration Management
python-dotenv==1.0.0

# Production WSGI Server
gunicorn==21.2.0

# HTTP Requests
requests==2.31.0

# Async Support and WebSockets (compatible version for Alpaca)
websockets==10.4

# System Utilities
psutil==5.9.5

# Development Dependencies (optional)
pytest==7.4.2
pytest-asyncio==0.21.1
